PP-4347 Moved state.js to config directory, renamed router.js -> routes.js and brought more things in line with self-service
