+++ app/controllers/charge_controller.js
@@ -14,7 +14,7 @@ const normalise = require('../services/normalise_charge')
 const chargeValidator = require('../utils/charge_validation_backend')
 const Charge = require('../models/charge')
 const Card = require('../models/card')
-const State = require('../models/state')
+const State = require('../../config/state')
 const paths = require('../paths')
 const CORRELATION_HEADER = require('../utils/correlation_header').CORRELATION_HEADER
 const {countries} = require('../services/countries')
